@@ -28,53 +28,52 @@
     "serve": "cd tools && npm run cli:serve"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.27",
-    "@angular-devkit/build-ng-packagr": "^0.803.27",
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/cli": "^8.3.27",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.901.12",
+    "@angular-devkit/build-ng-packagr": "~0.901.12",
+    "@angular/animations": "^9.1.12",
+    "@angular/cdk": "^9.2.4",
+    "@angular/cli": "^9.1.12",
+    "@angular/common": "^9.1.12",
+    "@angular/compiler": "^9.1.12",
+    "@angular/compiler-cli": "^9.1.12",
+    "@angular/core": "^9.1.12",
+    "@angular/flex-layout": "^9.0.0-beta.31",
+    "@angular/forms": "^9.1.12",
     "@angular/http": "^7.2.15",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/material": "^9.2.4",
+    "@angular/platform-browser": "^9.1.12",
+    "@angular/platform-browser-dynamic": "^9.1.12",
+    "@angular/router": "^9.1.12",
     "@compodoc/compodoc": "^1.1.11",
-    "@firestitch/checkboxgroup": "^7.1.10",
-    "@firestitch/common": "^7.0.5",
-    "@firestitch/drawer": "^8.0.2",
-    "@firestitch/example": "^8.0.0",
+    "@firestitch/checkboxgroup": "^8.1.5",
+    "@firestitch/common": "^8.0.5",
+    "@firestitch/drawer": "^9.1.5",
+    "@firestitch/example": "^9.1.0",
     "@firestitch/iframe": "^8.0.0",
-    "@firestitch/label": "^7.1.9",
-    "@firestitch/menu": "^7.0.11",
+    "@firestitch/label": "^8.2.5",
+    "@firestitch/menu": "^9.1.3",
     "@firestitch/message": "^8.1.0",
-    "@firestitch/selectbutton": "^8.0.0",
-    "codelyzer": "^5.2.1",
+    "@firestitch/selectbutton": "^8.0.1",
+    "codelyzer": "^5.1.2",
     "cpx": "^1.5.0",
     "hammerjs": "^2.0.8",
     "highlight.js": "^9.17.1",
     "lodash-es": "^4.17.15",
-    "ng-packagr": "^5.7.1",
-    "ngx-toastr": "^11.2.1",
+    "ng-packagr": "^9.0.0",
+    "ngx-highlightjs": "^4.0.2",
+    "ngx-toastr": "^12.1.0",
     "npm-run-all": "^4.1.5",
     "rxjs": "^6.5.3",
-    "tsickle": "^0.37.1",
     "tslib": "^1.10.0",
     "tslint": "^5.20.1",
-    "typescript": "^3.5.3",
-    "webpack": "^4.41.4",
-    "webpack-dev-server": "^3.10.1",
-    "zone.js": "^0.10.2"
+    "typescript": "~3.8.3",
+    "zone.js": "~0.10.2"
   },
   "engines": {
     "node": ">=6.0.0"
   },
   "peerDependencies": {
-    "@firestitch/selectbutton": ">=8.0.0"
+    "@firestitch/selectbutton": ">=8.0.0",
+    "tslib": ">=1.10.0"
   }
 }
